fix graph capture issue
with HPUHooks registeration, from autograd it sets
the stream to default during graph task execution.
In case of HPU graph capture it is done on different
user stream, becasue of this there will stream
switch during backward. This will cause HPU
graph capture to fail as the expectation is
during capture stream switch is not allowed. So fix
the graph capture to use default stream.

Change-Id: Ic2875093be949b8cb2c9b04e078e3d677e4ab356
Signed-off-by: Internal Developer <internal-developer@intel.com>