Fix test_pt2e_quant in PT2.6 (#433)
Since PT2.6, the torch.load function (called in a torch.export.load
function in this test case) has a weights_only parameter set to True by
default. This causes the issue when loading the model with custom
functions/classes. Such functions/classes must be added to the
safe_globals list in torch.serialization module beforhand so the model
can be loaded correctly.

For some unknown reason, issue with that tests caused the segfault in
test_op_randperm2 which was executed after test_pt2e_quant tests. The
real rootcause of the segfault will be investigated further in a ticket