Move inference env setup and teardown in tests to fixtures (#454)
*  Move inference env setup and teardown in tests to fixtures

Some tests require inference mode which is set up by calling appropriate
function. When the tests are finished, inference mode should be
disabled which is done by calling another function at the end of the
test. However, when the exception is raised during the test execution,
it may happen that the inference mode is not disabled what causes the
issues in next tests. To ensure the correct setup and teardown of
inference mode, it is moved to the pytest fixture. Thanks to that, the
teardown function will be always exectued.

---------

Co-authored-by: Internal Developer <internal-developer@intel.com>