Fix test_pt2e_quant in PT2.6 (#552)
Since PT2.6, the torch.load function (called in a torch.export.load
function in this test case) has a weights_only parameter set to True by
default. This causes the issue when loading the model with custom
functions/classes. Such functions/classes must be added to the
safe_globals list in torch.serialization module beforhand so the model
can be loaded correctly. It was fixed earlier in the PR 443, but in the
meanwhile new classes/functions were added to the tests which haven't
been added to the allow list.