Fix execution hang when using lazy_collective (#336)
When using the flags, compute and collective will be
part of the same execution thread. After launching the
compute, collectives are lunched. As part of compute
launch events are registered with callback function
to release the tensor reference held and to
decrease the recipe count. This callback is passed
to collective as well, this was causing the tensors
to be held till the collective completes. Due to this
memory was held for a longer time. In failure
case OOM happened during collective launch and was
waiting on the recipe count to decrease infinitely . This
never happens as the resource holder reference
was held by collective.
so, remove the linking of reference from compute
to collective.

Signed-off-by: Internal Developer <internal-developer@intel.com>
Co-authored-by: Internal Developer <internal-developer@intel.com>