FusedSDPA:Add support for fp32 softmax mode when attention inputs are BF16 (#726)
Add an API option "fp32"  for the existing 'softmax_mode' API parameter to run
softmax in FP32 precision. This is needed innorder to improve attention output
accuracy in certain topologies when inputs are in BF16. When The user sets
softmax_mode = "fp32", SDPA CGUID will set BGEMM1 output to fp32 and hence the
softmax input becomes fp32. If attention mask is present and is in bf16, it is
converted to FP32 in CGUID. The attention output still remains in BF16.

Change-Id: I36ce829dc8f02ea54be26e28b2a68720abe26c99

Signed-off-by: Internal Developer <internal-developer@intel.com>