Use explicit True/False value in a BuildConstant func for bool (#773)
Before that commit, when a constant TPC Kernel was called with a bool
precision via BuildConstant function, the param value could be any float.
This value was later used to fill in the output tensor. However, the
reinterpration of that value into True/False in the TPC Kernel is
sometimes incorrect, therefore this is moved now to the backend.