Fix one_hot shared layer validation (#777)
The backend of one_hot was recently removed and now the kernel is called
with the inputs passed without any modifications so the validator was
replaced with the default implementation. Additionally, the problem was
that in eager the validator was not executed at all. This was caused by
a custom op registration. To solve this, a validator call with custom
SharedMetaFunction call was added to the one_hot_forward function.