Fix BatchNorm1D single OP tests (#954)
*  Fix BatchNorm1D single OP tests

In the meta function for native_batch_norm OP, shape of the second and
third input was infered based on the input shapes of the running_mean and
the running_var tensors. However, sometimes the tensors are present,
but undefined what means they have shape 0. In that case, the output shape
should be infered from the channel size of the input tensor.

Additionally, with this commit some errors comming from the
validateOutputShapeNonDynamic are fixed, which didn't cause the test failure,
but still shouldn't be there.