Use fmin/fmax CGUIDs for fmax and fmin OPs. (#1526)
Current implementation of the fmax and fmin OPs is based on the
max_fwd and the min_fwd CGUIDs. Those kernels return NaN value when any
non-NaN and NaN is compared. However, according to the PyTorch documentation
for fmin/fmax OPs, when NaN and non-NaN is compared, the non-NaN should
be returned. To achieve that instead of max/min_fwd, fmin/fmax_fwd must
be used. As NaNs exist only for float dtypes, thus for integer dtypes
max/min_fwd still can be used.