Fix recipe cache calculation. (#1444)
We call reuse mem section, if same address is already used in
input, but it results in cache hit in cases where differnt address
is used in inputs, as recipe still has info that they reuse mem
section, but actually they don't.
It returns error of address conflict from synapse.

Signed-off-by: Internal Developer <internal-developer@intel.com>