Reland  Implement unique_dim eager (#1438)
unique_dim eager implementation is functional. The cause of the
revert was a dynamic shape single op test fail. This fail was caused by
a timeout of the test. The issue was reproduced locally. The test
execution time can be longer with certain argument combinations when
testing dynamic shapes. A PR was created:
https://gerrit.habana-labs.com/#/c/409376/
to adjust the test's input and shorten the execution time.

*The following was left unchanged from the initial commit*
The implementation is based on _unique eager and _unique2 eager

unique_dim is a non-inferable op. This means, that the output shape is not
known until the kernel is executed. This requires for some operations
(such as slicing and flipping) to be moved to frontend, as there is a
kernel execution present there. This kernel execution in frontend
produces an output tensor which is "back-padded" with zeros. In order
to aqcuire the tensor of the correct shape - which can be passed further
to backend - the zeros need to be sliced from the output tensor.

There is no torch::unique_dim() method available in python. However,
pytorch provides a universal torch.unique() method. This method is able
to delegate the call further down to _VF::unique_dim(), based on the
value of the provided 'dim' argument.
Moreover, torch::unique() method does not allow for empty input tensors.
To test an empty tensor input, a C++ test has been added, which is able
to call torch::unique_dim() directly.