Provide new manylinux docker images (#2636)
*  Provide new manylinux docker images

They're now based on PyTorch's manylinux images (built on top of
manylinux_2_28/AlmaLinux 8).

To allow easier migration of the images to dockerfiles-ci-cd in the
future, the existing CD entrypoint script was adopted (with small
adaptations to better suit our needs).

Python version selection should now be done by the HABANA_PYTHON_VERSION
environment variable and will require mounting a directory for a venv
from the host.

Relying on HABANA_PIP_VERSION was dropped as it is better to rely on the
latest available version. Some other tweaks include fixing the non-root
prompt, sudo permissions, sourcing the venv only once, providing users
with a better help and status messages.

An example invocation to get to an interactive bash shell is as follows:

docker run --rm -it -e HOST_USER=$USER -e HOST_UID=$(id -u) \
-e HOST_GID=$(id -g) -e _WORK=/ -e _STACK=/repos -e _BUILD=/builds \
-e HABANA_PYTHON_VERSION=3.11 \ -v ~/.ssh:/.ssh-host:ro
-v ~/trees/npu-stack:/repos -v ~/builds:/builds
-v /tmp/manylinux-venv-py3.11:/.venv image_name

NOTE: the venv directory must exist on the host and be writable by
$HOST_USER.

Other notes:
* IceCC threads used for rebuilds are no longer limited to just 1 thanks
  to replacing -m 0 with --no-remote
* the default GCC version is now 13 to stay in line with PT builders