Add support for diagonal_backward and diagonal_scatter in compile. (#3057)
*  Add support for diagonal_backward and diagonal_scatter in compile.

Before this commit, call to aten.diagonal_backward or aten.diagonal_scatter
in a compile mode resulted in a fallback to eager mode as no decomposition
nor native support was added for those OPs. With this change, both OPs
are supported in all modes.

Add unit tests for torch.diagonal_backward